Fix right image blurriness
This happens with Chrome on ChromeOS at odd-width resolutions, at least.
The right image is horizontally blurred, as if scaled by width+1px.
Removing the translate(-50%, 0) appears to fix the issue. However the page
will not be centered when the window is narrower than the content, it will be
left-aligned instead. I assume this is not a big deal because it is currently not
scrollable (as per <div class="demo" style="overflow:hidden;">), so it is meant
to be viewed in full at all times anyway.
I did not check this patch with other browsers or operating systems.